fix: restructered components
