+++ src/components/Dashboard/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import './Dashboard.css';
 import { Link } from 'react-router-dom';
+
+import styles from './index.css';
 
 function Dashboard() {
   return (
