+++ src/components/FetchButton/index.jsx
@@ -1,5 +1,7 @@
 import React, { useState } from 'react';
 
+import styles from './index.css';
+
 function FetchButton({ label, apiUrl }) {  // Компонент принимает название кнопки и URL API
   const [data, setData] = useState([]);
   const [loading, setLoading] = useState(false);
