+++ src/components/Home/index.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 
+import styles from './index.css';
+
 function Home() {
   const navigate = useNavigate(); // Хук для навигации
